requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
requests==2.32.3
pydantic==2.10.5
python-dotenv==1.0.1
rich==13.9.4
tenacity==9.0.0
unidiff==0.7.5
